update : minter,add merkle claim  and update deploy scripts
