+++ contracts/interfaces/IMinter.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-import {IVelo} from "./IVelo.sol";
+import {ITenex} from "./ITenex.sol";
 import {IVoter} from "./IVoter.sol";
 import {IVotingEscrow} from "./IVotingEscrow.sol";
 import {IRewardsDistributor} from "./IRewardsDistributor.sol";
@@ -19,8 +19,8 @@ interface IMinter {
     event Nudge(uint256 indexed _period, uint256 _oldRate, uint256 _newRate);
     event AcceptTeam(address indexed _newTeam);
 
-    /// @notice Interface of Velo.sol
-    function velo() external view returns (IVelo);
+    /// @notice Interface of Tenex.sol
+    function tenex() external view returns (ITenex);
 
     /// @notice Interface of Voter.sol
     function voter() external view returns (IVoter);
@@ -61,7 +61,7 @@ interface IMinter {
     /// @notice Tail emissions rate in basis points
     function tailEmissionRate() external view returns (uint256);
 
-    /// @notice Starting weekly emission of 15M VELO (VELO has 18 decimals)
+    /// @notice Starting weekly emission of 15M TENEX (TENEX has 18 decimals)
     function weekly() external view returns (uint256);
 
     /// @notice Timestamp of start of epoch that updatePeriod was last called in
@@ -81,6 +81,14 @@ interface IMinter {
     /// @notice Creates a request to change the current team's address
     /// @param _team Address of the new team to be chosen
     function setTeam(address _team) external;
+
+    /// @notice set redemptionReceiver address in tenex token 
+    function setRedemptionReceiver(address _receiver) external;
+
+    /// @notice set merkleclaim address in tenex token 
+    function setMerkleClaim(address _merkleclaim) external;
+
+
 
     /// @notice Accepts the request to replace the current team's address
     ///         with the requested one, present on variable pendingTeam
@@ -103,10 +111,10 @@ interface IMinter {
     function nudge() external;
 
     /// @notice Calculates rebases according to the formula
-    ///         weekly * (ve.totalSupply / velo.totalSupply) ^ 3 / 2
+    ///         weekly * (ve.totalSupply / tenex.totalSupply) ^ 3 / 2
     ///         Note that ve.totalSupply is the locked ve supply
-    ///         velo.totalSupply is the total ve supply minted
-    /// @param _minted Amount of VELO minted this epoch
+    ///         tenex.totalSupply is the total ve supply minted
+    /// @param _minted Amount of TENEX minted this epoch
     /// @return _growth Rebases
     function calculateGrowth(uint256 _minted) external view returns (uint256 _growth);
