+++ contracts/interfaces/ITenex.sol
@@ -14,13 +14,15 @@ interface ITenex is IERC20 {
     /// @return True if success
     function mint(address account, uint256 amount) external returns (bool);
 
+    function setRedemptionReceiver(address _receiver) external;
 
-    function claim(address, uint) external returns (bool);
+    function setMerkleClaim(address _merkleclaim) external;
+
+    function claim(address account, uint amount) external returns (bool);
 
     /// @notice Address of Minter.sol
     function minter() external view returns (address);
 
     /// @notice Address of MerkleClaim.sol
     function merkleClaim() external view returns (address);
-
 }
