+++ script/constants/Blast.json
@@ -9,14 +9,14 @@
         "0x4200000000000000000000000000000000000022",
         "0xD88957c98D65E9bee30304290f734847De09B990"
     ],
-    "poolsV2": [
+    "pools": [
         {
             "stable": false,
             "tokenA": "0x4200000000000000000000000000000000000022",
             "tokenB": "0xD88957c98D65E9bee30304290f734847De09B990"
         }
     ],
-    "poolsVeloV2": [
+    "poolsTenex": [
         {
             "stable": false,
             "token": "0x4200000000000000000000000000000000000022"
