+++ script/hardhat/DeployGaugesAndPools.ts
@@ -1,7 +1,7 @@
 import { getContractAt } from "./utils/helpers";
 import { PoolFactory, Voter } from "../../artifacts/types";
 import jsonConstants from "../constants/Blast.json";
-import deployedContracts from "../constants/output/VelodromeV2Output.json";
+import deployedContracts from "../constants/output/TenexOutput.json";
 
 async function main() {
   const factory = await getContractAt<PoolFactory>(
@@ -11,8 +11,8 @@ async function main() {
   const voter = await getContractAt<Voter>("Voter", deployedContracts.voter);
 
   // Deploy non-VELO pools and gauges
-  for (var i = 0; i < jsonConstants.poolsV2.length; i++) {
-    const { stable, tokenA, tokenB } = jsonConstants.poolsV2[i];
+  for (var i = 0; i < jsonConstants.pools.length; i++) {
+    const { stable, tokenA, tokenB } = jsonConstants.pools[i];
     await factory.functions["createPool(address,address,bool)"](
       tokenA,
       tokenB,
@@ -35,10 +35,10 @@ async function main() {
   }
 
   // Deploy VELO pools and gauges
-  for (var i = 0; i < jsonConstants.poolsVeloV2.length; i++) {
-    const [stable, token] = Object.values(jsonConstants.poolsVeloV2[i]);
+  for (var i = 0; i < jsonConstants.poolsTenex.length; i++) {
+    const [stable, token] = Object.values(jsonConstants.poolsTenex[i]);
     await factory.functions["createPool(address,address,bool)"](
-      deployedContracts.VELO,
+      deployedContracts.TENEX,
       token,
       stable,
       {
@@ -46,7 +46,7 @@ async function main() {
       }
     );
     let pool = await factory.functions["getPool(address,address,bool)"](
-      deployedContracts.VELO,
+      deployedContracts.TENEX,
       token,
       stable,
       {
@@ -60,6 +60,8 @@ async function main() {
     );
   }
 }
+
+
 
 main().catch((error) => {
   console.error(error);
