+++ test/Deploy.t.sol
@@ -113,7 +113,7 @@ contract TestDeploy is BaseTest {
         assertEq(deployVelodromeV2.VELO().minter(), address(deployVelodromeV2.minter()));
 
         assertEq(deployVelodromeV2.voter().minter(), address(deployVelodromeV2.minter()));
-        assertEq(address(deployVelodromeV2.minter().velo()), address(deployVelodromeV2.VELO()));
+        assertEq(address(deployVelodromeV2.minter().tenex()), address(deployVelodromeV2.VELO()));
         assertEq(address(deployVelodromeV2.minter().voter()), address(deployVelodromeV2.voter()));
         assertEq(address(deployVelodromeV2.minter().ve()), address(deployVelodromeV2.escrow()));
         assertEq(address(deployVelodromeV2.minter().rewardsDistributor()), address(deployVelodromeV2.distributor()));
